// demos/convex/http/convex/_generated/scheduled_functions.ts
// Generated by @neovex/codegen. Do not edit by hand.
import { makeMutationReference } from "convex/browser";
import type { Doc, Id } from "./dataModel";

type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export const scheduledFunctions = {
  messages: {
    send: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:send", "public"),
    scheduleSend: makeMutationReference<{
  "author": string;
  "body": string;
  "delayMs": number;
}, string>("messages:scheduleSend", "public"),
    sendAndSchedule: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendAndSchedule", "public")
  }
};

export const internalScheduledFunctions = {
  messages: {
    sendInternal: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendInternal", "internal")
  }
};